backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
redis==4.5.1
SQLAlchemy==2.0.7
Werkzeug==2.2.3
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
PyJWT==2.6.0
requests==2.28.2
python-dateutil==2.8.2
